unbox: add tool to inspect and unpack mailbox files
Exports from GMail arrive as a single Unix-style mbox file, which is tedious to
work with especially if you have a lot of messages. This tool helps with
listing out what is there, and unpacking it into individual files.

This does not do the harder part of unpacking attachments and deciphering the
various multipart/alternative presentation shenanigans, but it's a start.