get-the-gist: rewrite gist URLs to SSH format
When fetching, use the HTTP URL to avoid having to set up SSH transport.  This
works fine, since gists on the public instance are all readable without auth
(even those that are unpublished or "secret"). But store the SSH url in the
checkout, so that a human operator can push changes.